+++ lib/src/modules/favorites/widgets/filter_bar_favorite.dart
@@ -9,9 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class FilerBarFavorites extends StatelessWidget {
-  const FilerBarFavorites({
-    Key? key,
-  }) : super(key: key);
+  const FilerBarFavorites({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
