+++ lib/src/modules/favorites/widgets/product_card_vertical.dart
@@ -1,12 +1,12 @@
 import 'package:e_commerce/src/config/routes/coordinator.dart';
 import 'package:e_commerce/src/config/themes/my_colors.dart';
-import 'package:e_commerce/src/constants/my_icons.dart';
 import 'package:e_commerce/src/models/products_model.dart';
 import 'package:e_commerce/src/modules/favorites/logic/favorites_bloc.dart';
-import 'package:e_commerce/src/utils/utils.dart';
+import 'package:e_commerce/src/modules/product/widgets/display_label.dart';
+import 'package:e_commerce/src/modules/product/widgets/display_size_and_color.dart';
+import 'package:e_commerce/src/modules/product/widgets/price_product_widget.dart';
+import 'package:e_commerce/src/modules/product/widgets/review_star.dart';
 import 'package:e_commerce/src/widgets/button/button_add_bag.dart';
-import 'package:e_commerce/src/widgets/label/discount_label.dart';
-import 'package:e_commerce/src/widgets/label/new_label.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -35,15 +35,7 @@ class XProductCardFavoriteVertical extends StatelessWidget {
                           Row(
                             mainAxisAlignment: MainAxisAlignment.spaceBetween,
                             children: [
-                              Padding(
-                                padding: const EdgeInsets.all(8.0),
-                                child: data.newProduct == true
-                                    ? const NewLabel()
-                                    : (data.discount != 0
-                                        ? DisCountLabel(
-                                            number: data.discount.toString())
-                                        : const SizedBox.shrink()),
-                              ),
+                              XDisplayLabel(data: data),
                               BlocBuilder<FavoriteBloc, FavoriteState>(
                                 builder: (context, state) => Padding(
                                     padding: const EdgeInsets.all(8.0),
@@ -73,8 +65,6 @@ class XProductCardFavoriteVertical extends StatelessWidget {
   }
 
   Widget _belowCard() {
-    var numberStarActive = ((data.star).toInt() ~/ 5);
-    var numberStarNoActive = 5 - numberStarActive;
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
@@ -89,48 +79,7 @@ class XProductCardFavoriteVertical extends StatelessWidget {
             ),
           ),
         ),
-        Row(
-          children: [
-            SizedBox(
-              height: 20,
-              child: ListView.builder(
-                  scrollDirection: Axis.horizontal,
-                  shrinkWrap: true,
-                  itemBuilder: (context, index) {
-                    return SizedBox(
-                      width: 13,
-                      height: 12,
-                      child: Image.asset(MyIcons.activeStarIcon,
-                          fit: BoxFit.scaleDown),
-                    );
-                  },
-                  itemCount: numberStarActive),
-            ),
-            SizedBox(
-              height: 20,
-              child: ListView.builder(
-                  scrollDirection: Axis.horizontal,
-                  shrinkWrap: true,
-                  itemBuilder: (context, index) {
-                    return SizedBox(
-                      width: 13,
-                      height: 12,
-                      child:
-                          Image.asset(MyIcons.starIcon, fit: BoxFit.scaleDown),
-                    );
-                  },
-                  itemCount: numberStarNoActive),
-            ),
-            const SizedBox(width: 3),
-            Text(
-              "${data.star}",
-              style: const TextStyle(
-                  fontSize: 10,
-                  color: MyColors.colorGray,
-                  fontWeight: FontWeight.normal),
-            )
-          ],
-        ),
+        XReviewStar(numberStar: data.star),
         Text(
           data.name,
           style: const TextStyle(
@@ -148,72 +97,8 @@ class XProductCardFavoriteVertical extends StatelessWidget {
               height: 1,
               fontWeight: FontWeight.w600),
         ),
-        Row(
-          children: [
-            RichText(
-              text: TextSpan(
-                text: 'Color: ',
-                style: const TextStyle(
-                    fontSize: 11,
-                    height: 1,
-                    fontWeight: FontWeight.normal,
-                    color: MyColors.colorGray),
-                children: <TextSpan>[
-                  TextSpan(
-                      text: '${data.color}',
-                      style: const TextStyle(color: MyColors.colorBlack)),
-                ],
-              ),
-            ),
-            const SizedBox(
-              width: 30,
-            ),
-            RichText(
-              text: TextSpan(
-                text: 'Size: ',
-                style: const TextStyle(
-                    fontSize: 11,
-                    height: 1,
-                    fontWeight: FontWeight.normal,
-                    color: MyColors.colorGray),
-                children: <TextSpan>[
-                  TextSpan(
-                      text: '${data.size}',
-                      style: const TextStyle(color: MyColors.colorBlack)),
-                ],
-              ),
-            )
-          ],
-        ),
-        data.discount == 0.0
-            ? Text(
-                "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                style: const TextStyle(
-                    fontSize: 14,
-                    height: 1.42,
-                    color: MyColors.colorBlack,
-                    fontWeight: FontWeight.w600),
-              )
-            : RichText(
-                text: TextSpan(
-                  text: "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                  style: const TextStyle(
-                      fontSize: 14,
-                      height: 1.42,
-                      decoration: TextDecoration.lineThrough,
-                      color: MyColors.colorGray,
-                      fontWeight: FontWeight.w600),
-                  children: <TextSpan>[
-                    TextSpan(
-                        text:
-                            "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
-                        style: const TextStyle(
-                          color: MyColors.colorSaleHot,
-                          decoration: TextDecoration.none,
-                        )),
-                  ],
-                ),
-              ),
+        XDisplaySizeAndColor(data: data),
+        XPriceProductWidget(data: data)
       ],
     );
   }
@@ -235,15 +120,7 @@ class XProductCardFavoriteVertical extends StatelessWidget {
                   Row(
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     children: [
-                      Padding(
-                        padding: const EdgeInsets.all(8.0),
-                        child: data.newProduct == true
-                            ? const NewLabel()
-                            : (data.discount != 0
-                                ? DisCountLabel(
-                                    number: data.discount.toString())
-                                : const SizedBox.shrink()),
-                      ),
+                      XDisplayLabel(data: data),
                       BlocBuilder<FavoriteBloc, FavoriteState>(
                         builder: (context, state) => Padding(
                             padding: const EdgeInsets.all(8.0),
