+++ lib/src/modules/product_details/widgets/details.dart
@@ -1,10 +1,9 @@
 import 'package:e_commerce/src/config/themes/my_colors.dart';
-import 'package:e_commerce/src/constants/my_icons.dart';
 import 'package:e_commerce/src/models/products_model.dart';
-import 'package:e_commerce/src/utils/utils.dart';
+import 'package:e_commerce/src/modules/product/widgets/display_label.dart';
+import 'package:e_commerce/src/modules/product/widgets/price_product_widget.dart';
+import 'package:e_commerce/src/modules/product/widgets/review_star.dart';
 import 'package:e_commerce/src/widgets/button/button_primary.dart';
-import 'package:e_commerce/src/widgets/label/discount_label.dart';
-import 'package:e_commerce/src/widgets/label/new_label.dart';
 import 'package:flutter/material.dart';
 
 class XDetails extends StatelessWidget {
@@ -13,8 +12,6 @@ class XDetails extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var numberStarActive = ((data.star).toInt() ~/ 5);
-    var numberStarNoActive = 5 - numberStarActive;
     const String contentText =
         'Short dress in soft cotton jersey with decorative buttons down the front'
         ' and a wide, frill-trimmed square neckline with concealed elastication.'
@@ -35,35 +32,10 @@ class XDetails extends StatelessWidget {
                     color: MyColors.colorBlack,
                     fontWeight: FontWeight.w600),
               ),
-              data.discount == 0.0
-                  ? Text(
-                      "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                      style: const TextStyle(
-                          fontSize: 24,
-                          height: 1.42,
-                          color: MyColors.colorBlack,
-                          fontWeight: FontWeight.w600),
-                    )
-                  : RichText(
-                      text: TextSpan(
-                        text: "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                        style: const TextStyle(
-                            fontSize: 24,
-                            height: 1.42,
-                            decoration: TextDecoration.lineThrough,
-                            color: MyColors.colorGray,
-                            fontWeight: FontWeight.w600),
-                        children: <TextSpan>[
-                          TextSpan(
-                              text:
-                                  "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
-                              style: const TextStyle(
-                                color: MyColors.colorSaleHot,
-                                decoration: TextDecoration.none,
-                              )),
-                        ],
-                      ),
-                    ),
+              XPriceProductWidget(
+                data: data,
+                fontSize: 24,
+              )
             ],
           ),
           Row(
@@ -79,55 +51,10 @@ class XDetails extends StatelessWidget {
                           fontSize: 11,
                           fontWeight: FontWeight.normal,
                           height: 1)),
-                  Row(
-                    children: [
-                      SizedBox(
-                        height: 20,
-                        child: ListView.builder(
-                            scrollDirection: Axis.horizontal,
-                            shrinkWrap: true,
-                            itemBuilder: (context, index) {
-                              return SizedBox(
-                                width: 13,
-                                height: 12,
-                                child: Image.asset(MyIcons.activeStarIcon,
-                                    fit: BoxFit.scaleDown),
-                              );
-                            },
-                            itemCount: numberStarActive),
-                      ),
-                      SizedBox(
-                        height: 20,
-                        child: ListView.builder(
-                            scrollDirection: Axis.horizontal,
-                            shrinkWrap: true,
-                            itemBuilder: (context, index) {
-                              return SizedBox(
-                                width: 13,
-                                height: 12,
-                                child: Image.asset(MyIcons.starIcon,
-                                    fit: BoxFit.scaleDown),
-                              );
-                            },
-                            itemCount: numberStarNoActive),
-                      ),
-                      const SizedBox(width: 3),
-                      Text(
-                        "(${data.star})",
-                        style: const TextStyle(
-                            fontSize: 10,
-                            color: MyColors.colorGray,
-                            fontWeight: FontWeight.normal),
-                      )
-                    ],
-                  ),
+                  XReviewStar(numberStar: data.star)
                 ],
               ),
-              data.newProduct == true
-                  ? const NewLabel()
-                  : (data.discount != 0
-                      ? DisCountLabel(number: data.discount.toString())
-                      : const SizedBox.shrink()),
+              XDisplayLabel(data: data)
             ],
           ),
           const SizedBox(
