+++ lib/src/widgets/card/product_card_new.dart
@@ -1,10 +1,10 @@
 import 'package:e_commerce/src/config/routes/coordinator.dart';
 import 'package:e_commerce/src/config/themes/my_colors.dart';
-import 'package:e_commerce/src/constants/my_icons.dart';
 import 'package:e_commerce/src/models/products_model.dart';
-import 'package:e_commerce/src/utils/utils.dart';
+import 'package:e_commerce/src/modules/product/widgets/display_label.dart';
+import 'package:e_commerce/src/modules/product/widgets/price_product_widget.dart';
+import 'package:e_commerce/src/modules/product/widgets/review_star.dart';
 import 'package:e_commerce/src/widgets/button/button_add_favorite.dart';
-import 'package:e_commerce/src/widgets/label/new_label.dart';
 import 'package:flutter/material.dart';
 
 class XProductCardNew extends StatelessWidget {
@@ -26,13 +26,9 @@ class XProductCardNew extends StatelessWidget {
               child: Column(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    const Align(
-                      alignment: Alignment.topLeft,
-                      child: Padding(
-                        padding: EdgeInsets.all(8.0),
-                        child: NewLabel(),
-                      ),
-                    ),
+                    Align(
+                        alignment: Alignment.topLeft,
+                        child: XDisplayLabel(data: data)),
                     Align(
                       alignment: Alignment.bottomRight,
                       child: XButtonAddToFavorite(
@@ -48,8 +44,6 @@ class XProductCardNew extends StatelessWidget {
   }
 
   Widget _belowCard() {
-    var numberStarActive = ((data.star).toInt() ~/ 5);
-    var numberStarNoActive = 5 - numberStarActive;
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
@@ -64,48 +58,7 @@ class XProductCardNew extends StatelessWidget {
             ),
           ),
         ),
-        Row(
-          children: [
-            SizedBox(
-              height: 20,
-              child: ListView.builder(
-                  scrollDirection: Axis.horizontal,
-                  shrinkWrap: true,
-                  itemBuilder: (context, index) {
-                    return SizedBox(
-                      width: 13,
-                      height: 12,
-                      child: Image.asset(MyIcons.activeStarIcon,
-                          fit: BoxFit.scaleDown),
-                    );
-                  },
-                  itemCount: numberStarActive),
-            ),
-            SizedBox(
-              height: 20,
-              child: ListView.builder(
-                  scrollDirection: Axis.horizontal,
-                  shrinkWrap: true,
-                  itemBuilder: (context, index) {
-                    return SizedBox(
-                      width: 13,
-                      height: 12,
-                      child:
-                          Image.asset(MyIcons.starIcon, fit: BoxFit.scaleDown),
-                    );
-                  },
-                  itemCount: numberStarNoActive),
-            ),
-            const SizedBox(width: 3),
-            Text(
-              "${data.star}",
-              style: const TextStyle(
-                  fontSize: 10,
-                  color: MyColors.colorGray,
-                  fontWeight: FontWeight.normal),
-            )
-          ],
-        ),
+        XReviewStar(numberStar: data.star),
         Text(
           data.name,
           style: const TextStyle(
@@ -122,14 +75,9 @@ class XProductCardNew extends StatelessWidget {
               height: 1,
               fontWeight: FontWeight.w600),
         ),
-        Text(
-          "${XUtils.formatPrice(data.originalPrice)}\$ ",
-          style: const TextStyle(
-              fontSize: 14,
-              height: 1.42,
-              color: MyColors.colorBlack,
-              fontWeight: FontWeight.w600),
-        ),
+        XPriceProductWidget(
+          data: data,
+        )
       ],
     );
   }
